requirements.txt
sqlalchemy
fastapi
psycopg2-binary
dotenv
uvicorn

# Auth
python-jose[cryptography]
passlib[bcrypt]
python-multipart
